requirements.txt
PyMuPDF==1.23.22
numpy==1.24.3
Pillow==10.2.0 